code review adjusts
