+++ docker-compose.yml
@@ -6,8 +6,6 @@ services:
     image: mariadb:10.5.9
     env_file:
       - .env
-    ports:
-      - "3306:3306"
     volumes:
       - csdl-db-volume:/var/lib/mysql
     command: mysqld --default-authentication-plugin=mysql_native_password
